Refactor generator package layout
