+++ cgmes_generator/parser/__init__.py
@@ -1,0 +1,12 @@
+"""Parsing utilities for CGMES XMI files."""
+
+from .loader import load_xmi
+from .multiplicity import mult_from_elem, ptype
+from .xmi import parse_xmi
+
+__all__ = [
+    "load_xmi",
+    "mult_from_elem",
+    "ptype",
+    "parse_xmi",
+]
